LayoutTests/canvas/philip/tests/2d.imageData.create2.tiny-expected.txt

< [index] >
2d.imageData.create2.tiny
createImageData(sw, sh) works for sizes smaller than one pixel
References: 2d.imageData.create2.size 2d.imageData.one Bugzilla
Actual output:
Passed
